Use external file for backup data.
Instead of typing the individual disk/repo passwords each time, save
them to a root-accessible-only file and source it in the backup/restore
scripts. This allows us to source them as root (required as the
perms/ownership are root-only for the data file) without having to run
restic itself as root, which prevents the restored directory from having
root as the owner.