Create pyutils.
This will contain some common utilities when creating basic tools in
python. The first such utility is an easy interface for calling into the
shell, which should prove very useful in cases where we want to call
complex shell pipelines but don't want to have to deal with basically
anything else in bash.