Add grep (-g) option to open tool.
This option allows for searching file contents, rather than file names.
The tool will then open the file as normal, either via nvim or xdg-open.

In the future this may be tweaked to open the file at the matched line,
assuming the file is to be opened by nvim and nvim has an option for
opening a file at a particular line.