Automatically determine the "main" repo branch.
The old default used to be "master" but it seems this is slowly shifting
to "main". We therefore attempt to automatically determine which branch
is the principle branch of the repo, rather than hardcoding "master".
Since it'd be a pain in the butt to update all my old repos, we assume
that exactly one of those two branch names exists in the repo.