Remove .mozilla from backup for now.
Seems the name of the profile can change between computers so perhaps a
manual backup of the specific files I want to keep (e.g. places.sqlite)
is more appropriate for when I switch between computers. Also .mozilla
is larger than I thought and seems to change a lot so we avoid that
churn for now.