Transition away from fm/ "monorepo".
We now simply use the home directory directly. This does result in
slightly less portability, as the portions that are backed up are a
subset of the full home directory, but makes day-to-day usage more
convenient.

As part of this migration, the standard XDG directories were renamed in
the associated user-dirs.dirs config.