Add systemd user service for transcribe_watcher.
This allows for the transcribe_watcher script to run automatically upon
user login. We had to make a couple of minor changes to
transcribe_watcher in order for this to work properly in systemd:

* Use full path to transcribe script, as it seems as though systemd does
  not include that PATH change in its environment.
* Include an additional filter on the watched files to exclude .trashed-
  versions that syncthing automatically creates when deleting a file
  from the directory (had to resort to grep here because inotifywait
  does not allow specifying both --include and --exclude patterns).

Additionally, the service itself needs to start before
syncthing.service, since the directory we're watching is a syncthing
sync directory, and we want to make sure we start the watcher before
syncthing has a chance to sync any new files.