Overhaul backup scripts.
This commit introduces a new `rbackup` script that performs a backup to
a restic repository. The script expects a bash config file as an
argument, whose job is to provide BACKUP_INCLUDE and optionally
BACKUP_EXCLUDE and RESTIC_PASSWORD variables. This allows to more easily
say "back up these things (but not these other things) to this repo",
without having to create a homegrown script for every backup scenario.
For example, it's now much easier to backup to a repo on an external
drive once it's mounted.

The existing `homebackup` and `bwbackup` scripts are kept for
convenience, as those are the most common backup scenarios, but
`homebackup` now simply calls out to `rbackup` to do the actual local
and remote backups.

`bwbackup` is largely unchanged in this respect though, as we use restic
slightly differently there (i.e. piping bw vault data directly into
restic), but a future commit might allow for similar parameterization.