WIP: create notification server for notify-send.
Trying to get systemd user services to send notifications sucks so it
might be easier to just start up a simple python http server that does
it via desktop autoload and query that from systemd when I want to send
a notification (e.g. when a new voice transcription arrives).