Reintroduce simplified `notes` script.
This version is much simpler and more streamlined, basically just
searching through my notes directory for file names and their tags and
presenting those lines to `fzf` so that I can open the selected file
with `helix`.

Also change Alt-w in kitty to use this new script instead of opening the
wiki index in neovim.