Add script + udev rules for power connect/disconnect.
Now when my laptop is connected to mains power, it automatically
switches to "performance" power mode, and when it's disconnected it
instead switches to "power saver". I had tried to do something like this
previously but never managed to get it to work, credit to [1] for
showing me the proper way of doing it.

[1]: https://dataswamp.org/~solene/2025-05-31-linux-killswitch-on-power-disconnect.html